Add Bloc observer and setup service locator in main.dart file
This commit adds a Bloc observer to the main function and sets up the service locator for the API service and map in the main.dart file. The SimpleBlocObserver is used to observe the Bloc events and the setupServiceLocatorApiService() and setupServiceLocatorMap() functions are called to set up the service locator.