Add MapState class to MapCubit feature
This commit adds the MapState class to the MapCubit feature. The MapState class contains four subclasses: MapInitial, MapSuccess, MapLoading, and MapFailure. These subclasses represent the different states of the MapCubit feature and are used to update the UI accordingly. The MapSuccess subclass also contains a RoutesModel object to pass data to the UI.