Add UseCase and RoutesUseCase classes
Added UseCase and RoutesUseCase abstract classes to provide a common structure for implementing use cases in the application. These classes will be used to handle business logic and return either a success or failure response. Also added a NoParam class to handle use cases that do not require any parameters.